Update more test cases..
